move user impl to top-level user package
